requirements.txt
# "python-3.10.10-embed-amd64\python" "python-3.10.10-embed-amd64\Scripts\pip.exe" install -r requirements.txt
# "python-3.10.10-embed-amd64\python" "python-3.10.10-embed-amd64\Scripts\pip.exe" list
PyQt5
opencv-python
pyinstaller
ffmpeg-python